fix: BigQuery destination: wait for table migrations to complete (#5544)
This adds additional wait blocks to the BigQuery destination migration code to ensure that we do not return from `Migrate` before the tables are confirmed to be either created or matching the expected schema, depending on the case. This helps prevent an issue where BigQuery returns a 404 when a write immediately follows the creation of the table (as is usually the case when you run `cloudquery sync` for the first time).